fix(playerbot): Phase 4C - Fix PlayerbotCommands and remaining API errors
Fixed 15+ additional compilation errors:

PlayerbotCommands fixes:
- RBAC_PERM_COMMAND_GM_NOTIFY → RBAC_PERM_COMMAND_GMNOTIFY
- Trinity::ChatCommands::Optional → std::optional<T>
- Group::GetFirstMember() → range-based for with GetMembers()
- Added MotionMaster.h include
- Removed deprecated ChrClassesEntry::IsRaceValid check

Other file fixes:
- GroupFormationManager: Cast GetPrimarySpecialization() to uint32
- BotStatePersistence: Added Bag.h include, fixed Item durability API
- QuestPathfinder: Created stub QuestHubDatabase.h header
- UnifiedInterruptSystem: Fixed BotInterruptAssignment struct fields
- UnifiedInterruptSystem: Cast uint8 to SpellEffIndex for GetEffect()

Files modified: 8
Errors reduced: ~28 → ~15 (est.)

Related: Phase 4C (Test Compilation)

+++ src/modules/Playerbot/Movement/GroupFormationManager.cpp
@@ -257,7 +257,7 @@ namespace Playerbot
             return BotRole::UTILITY;
 
         uint8 classId = bot->GetClass();
-        uint32 specId = bot->GetPrimarySpecialization();
+        uint32 specId = static_cast<uint32>(bot->GetPrimarySpecialization());
 
         // Determine role based on class and specialization
         switch (classId)
